Lazy test of else if statements
Removed lines of iprintln that only applies to my own server.

I don't think you actually can use several "else if" statements like this, but it compiles and doesn't crash so far but I'm not sure if it's working yet due to nobody joining on my 10+ round games today.

Even if it is working, this method is pretty bad.
Leaving it here for now since it atleast doesn't harm anything.