Nerf points gained & buff points lost on down/bleedout
Setting any of the PlayerPoints variables below their original value will result in minus points but it works as intended, when you normally shoot a bullet into a zombie, you gain +10 points.

Same thing happens if we set the value to 5 but when we shoot the zombie it will say +10 and -5 points at the same time.
So it works as it should, sadly we can't make it just say +5 instead of +10 -5

Punish players that are fucking around and getting downed 24/7 by doubling the amount of points they lose from default values.
Also punish players who doesn't revive harder.